Refactor into hooks and add tests
